requirements.txt
Flask
zappa
pydantic
requests
boto3
pyhumps